Make sure packaging and dissassembly actually works
It is harder to split a string representing a number into correctly
sized chunks to do math on than to do the same thing with a long
binary string.

Also make the tests more robust in the sense that the results no
longer need to be manually checked.

Next step now is reading e, d, and n form a standard rsa key file?