Add script for merging file cunks
For now a naive approach that assumes that all of the chunks are alone
in one directory, and that you don't have to consider sorting order.

Another limmitation is that you have to specify the dersired output path,
perhaps it would be better if the util could simply be pointed to a directory
with chunks and place the resulting merged file there.