Kafka KeyValue API: Add service for cryptography
Useful if you don't want to store the keys and/or values (or anything
representable as a byte array really) unencrypted. Note that this relies
on AES, because after lots of reading tonight, the state of the art still
seems to be that asymmetric crypto is only for tiny values (typically
less than the typical key size of maybe a couple of kilobytes), and that
you should use symmetric crypto for bigger collections of bytes, like
files or continuous http streams. And AES is still generally recommended.